:sparkles: finished seting up mobx
