+++ src/App.tsx
@@ -3,7 +3,8 @@ import React, { FunctionComponent } from 'react';
 import { ThemeProvider } from './lib/styled-components';
 import { theme } from './styles/theme';
 import { AppContainer } from './navigation';
-import { StoreProvider, rootStore } from './modules/rootStore';
+import { rootStore } from './modules/rootStore';
+import { StoreProvider } from './modules/provider';
 
 export const App: FunctionComponent = () => (
   <ThemeProvider theme={theme}>
