+++ src/modules/provider.ts
@@ -1,0 +1,6 @@
+import { createContext } from 'react';
+
+import { rootStore } from './rootStore';
+
+export const StoreContext = createContext(rootStore);
+export const StoreProvider = StoreContext.Provider;
