+++ src/modules/rootModel.ts
@@ -1,7 +1,5 @@
-import { types } from 'mobx-state-tree';
+import { User } from './user';
 
-import { UserModel } from './user';
-
-export const RootStore = types.model({
-  user: UserModel,
-});
+export const RootModel = {
+  user: User,
+};
