+++ src/modules/types.ts
@@ -1,0 +1,4 @@
+import { rootStore } from './rootStore';
+
+export type RootStoreType = typeof rootStore;
+export type SelectorFunction<T> = (store: RootStoreType) => T;
