+++ src/modules/user/model.ts
@@ -1,0 +1,5 @@
+import { types } from 'mobx-state-tree';
+
+export const userModel = {
+  name: types.optional(types.string, ''),
+};
