+++ src/modules/user/selectors.ts
@@ -1,3 +1,3 @@
-import { RootStoreType } from '../rootStore';
+import { RootStoreType } from '../types';
 
 export const userSelector = (store: RootStoreType) => store.user;
