+++ src/modules/workouts/index.ts
@@ -1,12 +1,3 @@
 import { types } from 'mobx-state-tree';
 
-const Todo = types
-  .model({
-    name: types.optional(types.string, ''),
-    done: types.optional(types.boolean, false),
-  })
-  .actions(self => ({
-    setName(newName) {
-      self.name = newName;
-    },
-  }));
+export const Workout = types.model();
