+++ src/pages/Home/Home.tsx
@@ -4,19 +4,17 @@ import { observer } from 'mobx-react-lite';
 import { TextInput } from 'react-native-gesture-handler';
 
 import styled from '../../lib/styled-components';
-import { useStore } from '../../modules/rootStore';
+import { useStore } from '../../utils/hooks/useStore';
 import { userSelector } from '../../modules/user/selectors';
 
 export const Home = observer(() => {
   const user = useStore(userSelector);
   const [userName, setUserName] = useState();
-  console.log(user);
 
   return (
     <Container>
-      <AppTitle>Workout app</AppTitle>
       <TextInput onChangeText={setUserName} placeholder="userName" value={userName} />
-      <AppTitle>{user.name}</AppTitle>
+      <StyledText>{user.name}</StyledText>
       <Button
         title="Valider"
         onPress={() => {
@@ -27,7 +25,7 @@ export const Home = observer(() => {
   );
 });
 
-const AppTitle = styled(Text)`
+const StyledText = styled(Text)`
   font-size: 24;
   text-align: center;
   margin: ${({ theme }): number => theme.margin.x3}px;
